Relocate utility functions
