+++ src/utils.h
@@ -1,0 +1,7 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+char* rfind(char* s, char target);
+int   mystrcmp(const char* s1, const char* s2);
+
+#endif // UTILS_H
